requirements.txt
pyephem==3.7.6.0
requests==2.18.4
